refactor: 컴포넌트별 디렉토리 구조
