+++ src/home/components/main/ComingSoon.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import { performanceApi, ComingSoonCard } from '../api/performanceApi';
-import { useTimeConversion } from '../../hooks/useTimeConversion';
-import '../styles/ComingSoon.css';
+import { performanceApi, ComingSoonCard } from '../../api/performanceApi';
+import { useTimeConversion } from '../../../hooks/useTimeConversion';
+import '../../styles/ComingSoon.css';
 
 // 이미지 로딩 상태를 관리하는 컴포넌트
 const ImageWithSkeleton: React.FC<{ src: string; alt: string }> = ({ src, alt }) => {
