+++ src/home/pages/Performance/PerformanceDetailPage.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import Layout from '../../../components/layout/Layout';
-import PerformanceDetail from '../../components/PerformanceDetail';
+import PerformanceDetail from '../../components/performance/PerformanceDetail';
 import { useScrollToTop } from '../../../hooks/useScrollToTop';
 
 const PerformanceDetailPage: React.FC = () => {
